+++ tests/parse/test_spatyper.py
@@ -10,10 +10,14 @@ def test_parse_spatyper_results(saureus_spatyper_path):
 
     # test parsing the output of saureus.
     result = parse_spatyper_results(saureus_spatyper_path)
-    expected_results = {
-        "sequence_name": "contig_1",
-        "repeats": "15-12-16-02-16-02-25-17-24",
-        "type": "t021"
+    expected_spatyper = {
+        "type": "spatype",
+        "software": "spatyper",
+        "result": {
+            "sequence_name": "contig_1",
+            "repeats": "15-12-16-02-16-02-25-17-24",
+            "type": "t021"
+        }
     }
     # check if data matches
-    assert expected_results == result[1].result.model_dump() # does this work?
+    assert expected_spatyper == result.model_dump()